Get rid of broken architecture that was a dead-end.
Stuff was refactored out at some point to different command handler
objects, but at some point dependencies between various handlers were
created (the design didn't handle that very well)

Also in principle there are only two commands, or now 'requests', one to
render a video or a frame (which can be an image, bitmap, etc.).

Attributes for those two requests can now be set, and should only
'amend' the main request, for example with extra annotations, or raw
versions of the image as output, etc.

The client-server feature hasn't been tested yet, but is is affected by
this refactoring. The CLI and UI code seems to still work at first
glance.

This refactoring also got rid of lots of parameters passing, which
introduced more specific coupling, now in some cases we just pass around
the original messages in it's entirely.

The refactoring first simplifies things, makes the starcry class
slightly bigger, but now we can refactoring things into helper classes
in a better way.

+++ src/data/job.hpp
@@ -25,7 +25,7 @@ struct job {
   uint32_t offset_y;
   uint32_t canvas_w;
   uint32_t canvas_h;
-  uint32_t job_number;
+  size_t job_number;
   uint32_t frame_number;
   bool rendered;
   bool last_frame;
